Add GNU GPL License
Created a license file for the GNU GPL v3 License. I want the software to be open source and free-to-use, but I don't want people taking the code and making a closed-source, commercial product. The whole point of this project was to provide organizations a free, open-source alternative to the many scheduler applications available that all have a high price tag.